+++ src/Data/Fields/Schema/FieldGeneric.php
@@ -11,7 +11,7 @@ class FieldGeneric extends BaseData implements FieldInterface
     /**
      * @param mixed[] $data
      */
-    public function __construct(protected string $fieldKey, array $data)
+    public function __construct(protected string $fieldKey, array $data = [])
     {
         $this->data = $data;
     }
@@ -77,5 +77,47 @@ class FieldGeneric extends BaseData implements FieldInterface
     public function tooltip(): bool
     {
         return $this->getBoolean("tooltip");
+    }
+
+    public function setPos(int $pos): static
+    {
+        $this->set("pos", $pos);
+        return $this;
+    }
+
+    public function setDisplayName(string $displayName): static
+    {
+        $this->set("display_name", $displayName);
+        return $this;
+    }
+
+    public function setRequired(bool $required = true): static
+    {
+        $this->set("required", $required);
+        return $this;
+    }
+
+    public function setTranslatable(bool $translatable = true): static
+    {
+        $this->set("translatable", $translatable);
+        return $this;
+    }
+
+    public function setNoTranslate(bool $noTranslate = true): static
+    {
+        $this->set("no_translate", $noTranslate);
+        return $this;
+    }
+
+    public function setDescription(string $description): static
+    {
+        $this->set("description", $description);
+        return $this;
+    }
+
+    public function setTooltip(bool $tooltip = true): static
+    {
+        $this->set("tooltip", $tooltip);
+        return $this;
     }
 }