+++ src/Data/Fields/Schema/FieldRichtext.php
@@ -6,6 +6,17 @@ namespace Storyblok\ManagementApi\Data\Fields\Schema;
 
 class FieldRichtext extends FieldGeneric
 {
+    public const TYPE = "richtext";
+
+    /**
+     * @param mixed[] $data
+     */
+    public function __construct(string $key, array $data = [])
+    {
+        $data["type"] = self::TYPE;
+        parent::__construct($key, $data);
+    }
+
     /**
      * @return array<mixed>
      */
@@ -14,8 +25,23 @@ class FieldRichtext extends FieldGeneric
         return $this->getArray("toolbar");
     }
 
+    /**
+     * @param string[] $toolbar
+     */
+    public function setToolbar(array $toolbar): static
+    {
+        $this->set("toolbar", $toolbar);
+        return $this;
+    }
+
     public function restrictComponents(): bool
     {
         return $this->getBoolean("restrict_components");
+    }
+
+    public function setRestrictComponents(bool $restrict = true): static
+    {
+        $this->set("restrict_components", $restrict);
+        return $this;
     }
 }